added security code and checks to location, events and sports and partly memberships. Removed public and secure controllers

+++ backend/src/main/java/nl/maastrichtuniversity/myusc/service/SportService.java
@@ -16,20 +16,21 @@ public class SportService {
     }
 
     public void addSport(SportDto sportDto) {
-        sportRepository.findByName(sportDto.getName())
+        sportRepository.findByNameIgnoreCase(sportDto.getName())
                 .ifPresent(s -> {
                     throw new RuntimeException("Sport with name " + sportDto.getName() + " already exists");
                 });
         Sport sport = new Sport();
-        sport.setName(sportDto.getName().toLowerCase());
+        sport.setName(sportDto.getName());
+        sport.setDescription(sportDto.getDescription());
         sport.setSportType(sportDto.getSportType());
         sportRepository.save(sport);
     }
 
     public void deleteSport(Long id) {
-//        if (!sportRepository.existsById(id)) {
-//            throw new RuntimeException("Sport with id " + id + " does not exist");
-//        }
+        if (!sportRepository.existsById(id)) {
+            throw new RuntimeException("Sport with id " + id + " does not exist");
+        }
         sportRepository.deleteById(id);
     }
 }
